refactor: cleanup old Zeitgrenzen-related stuff
RISDEV-7420
