+++ frontend/src/views/amending-law/articles/editor/multiple-mods/AmendingLawArticleEditorMultiMod.view.vue
@@ -8,7 +8,7 @@ import RisLawPreview from "@/components/RisLawPreview.vue"
 import { useDokumentExpressionEliPathParameter } from "@/composables/useDokumentExpressionEliPathParameter"
 import { useNormRenderHtml } from "@/composables/useNormRender"
 import { useSentryTraceId } from "@/composables/useSentryTraceId"
-import { useTemporalData } from "@/composables/useTemporalData"
+import { useTemporalData } from "@/views/amending-law/articles/editor/useTemporalData"
 import { useErrorToast } from "@/lib/errorToast"
 import Button from "primevue/button"
 import { useToast } from "primevue/usetoast"
