+++ frontend/src/views/amending-law/articles/editor/temporalDataResponse.ts
